Move custom image methods to own file, fix response
