+++ src/prelude.rs
@@ -1,7 +1,7 @@
 //! Wildcard glob this module to have access to all commonly used items.
 
 pub use api::{
-    Account, Action, Certificate, Domain, DomainRecord, Droplet, FloatingIp, Image, LoadBalancer,
+    Account, Action, Certificate, CustomImage, Domain, DomainRecord, Droplet, FloatingIp, Image, LoadBalancer,
     Region, Size, Snapshot, SshKey, Tag, Volume,
 };
 pub use request::Executable;
