+++ tests/image_action.rs
@@ -96,33 +96,3 @@ fn convert_produces_correct_request() {
     })
     );
 }
-
-#[test]
-fn create_produces_correct_request() {
-    before();
-
-    let correct_url = "https://api.digitalocean.com/v2/images";
-
-    let req: Request<Create, ()> = Image::create(
-        "test",
-        "http://foo.bar/baz.raw",
-        "ams3",
-        "archlinux",
-        "test image",
-        vec!["arch", "small"],
-    );
-    info!("{:#?}", req);
-
-    assert_eq!(req.url().as_str(), correct_url);
-    assert_eq!(
-        *req.body(),
-        json!({
-        "name": "test",
-        "url": "http://foo.bar/baz.raw",
-        "region": "ams3",
-        "distribution": "archlinux",
-        "description": "test image",
-        "tags": vec!["arch", "small"]
-    })
-    );
-}
